Fix #9 SocketBuilderExtensions.ConfigureMulticastOptions throws System.ArgumentException for IPv6 sockets

+++ OOs.Common.Net/Sockets/SocketBuilderExtensions.cs
@@ -66,7 +66,7 @@ public static class SocketBuilderExtensions
             }
             else
             {
-                socket.SetSocketOption(optionLevel, MulticastInterface, mcintAddress.ScopeId);
+                socket.SetSocketOption(optionLevel, MulticastInterface, (int)mcintAddress.ScopeId);
             }
         }
